rename ruby to Ruby
