+++ packages/quicktype-core/src/language/Ruby/index.ts
