+++ packages/quicktype-core/src/language/Ruby/keywords.ts
